relocate requirement tests
